requirements.txt
langchain>=0.1.0
langchain-google-genai>=1.0.0
langchain-community>=0.0.10
faiss-cpu>=1.7.4
pypdf>=3.17.0
pdfplumber>=0.10.0
python-dotenv>=1.0.0
requests>=2.31.0
streamlit>=1.28.0
langsmith>=0.1.0
pandas>=2.1.0
numpy>=1.24.0
python-multipart>=0.0.6